Finish User Module

+++ app.py
@@ -201,8 +201,11 @@ class Customer(object):
 
 @app.route("/")
 def index():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     # With trx data
-    return User.is_logged_in(render_template('dashboard/index.html'))
+    return render_template('dashboard/index.html')
 
 @app.route("/login/")
 def login():
@@ -211,6 +214,9 @@ def login():
 
 @app.route("/logout/")
 def logout():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     # With trx data
     return User.logout(redirect(url_for('login')))
 
@@ -218,22 +224,37 @@ def logout():
 
 @app.route("/transaction/")
 def transaction():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('transaction/index.html')
 
 @app.route("/transaction/add/")
 def transaction_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return IV_Transaction.add()
 
 @app.route("/transaction/read/<id>/")
 def transaction_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return IV_Transaction.read(id)
 
 @app.route("/transaction/update/")
 def transaction_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return IV_Transaction.update()
 
 @app.route("/transaction/delete/")
 def transaction_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return IV_Transaction.delete()
 
 # End Inventory Transaction Route
@@ -241,23 +262,38 @@ def transaction_delete():
 
 @app.route("/account_receivable/")
 def account_receivable():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('account_receivable/index.html')
 
 @app.route("/account_receivable/add/")
 def account_receivable_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return AR_Transaction.add()
 
 @app.route("/account_receivable/read/<id>/")
 def account_receivable_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('account_receivable/detail.html')
     # return AR_Transaction.read(id)
 
 @app.route("/account_receivable/update/")
 def account_receivable_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return AR_Transaction.update()
 
 @app.route("/account_receivable/delete/")
 def account_receivable_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return AR_Transaction.delete()
 
 # End AR Transaction Route
@@ -265,23 +301,38 @@ def account_receivable_delete():
 
 @app.route("/account_payable/")
 def account_payable():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('account_payable/index.html')
 
 @app.route("/account_payable/add/")
 def account_payable_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return AP_Transaction.add()
 
 @app.route("/account_payable/read/<id>/")
 def account_payable_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('account_payable/detail.html')
     # return AP_Transaction.read(id)
 
 @app.route("/account_payable/update/")
 def account_payable_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return AP_Transaction.update()
 
 @app.route("/account_payable/delete/")
 def account_payable_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return AP_Transaction.delete()
 
 # End AP Transaction Route
@@ -289,22 +340,37 @@ def account_payable_delete():
 
 @app.route("/purchase_order/")
 def purchase_order():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('')
 
 @app.route("/purchase_order/add/")
 def purchase_order_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PO_Transaction.add()
 
 @app.route("/purchase_order/read/<id>/")
 def purchase_order_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PO_Transaction.read(id)
 
 @app.route("/purchase_order/update/")
 def purchase_order_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PO_Transaction.update()
 
 @app.route("/purchase_order/delete/")
 def purchase_order_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PO_Transaction.delete()
 
 # End PO Transaction Route
@@ -312,22 +378,37 @@ def purchase_order_delete():
 
 @app.route("/purchase_receiving/")
 def purchase_receiving():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('')
 
 @app.route("/purchase_receiving/add/")
 def purchase_receiving_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PR_Transaction.add()
 
 @app.route("/purchase_receiving/read/<id>/")
 def purchase_receiving_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PR_Transaction.read(id)
 
 @app.route("/purchase_receiving/update/")
 def purchase_receiving_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PR_Transaction.update()
 
 @app.route("/purchase_receiving/delete/")
 def purchase_receiving_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return PR_Transaction.delete()
 
 # End Purchase Receiving (GRN) Route
@@ -335,23 +416,38 @@ def purchase_receiving_delete():
 
 @app.route("/inventory/")
 def inventory():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('inventory/index.html')
 
 @app.route("/inventory/add/")
 def inventory_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Inventory.add()
 
 @app.route("/inventory/read/<id>/")
 def inventory_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('inventory/detail.html')
     # return Inventory.read(id)
 
 @app.route("/inventory/update/")
 def inventory_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Inventory.update()
 
 @app.route("/inventory/delete/")
 def inventory_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Inventory.delete()
 
 # End Inventory Route
@@ -359,23 +455,38 @@ def inventory_delete():
 
 @app.route("/supplier/")
 def supplier():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('supplier/index.html')
 
 @app.route("/supplier/add/")
 def supplier_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Supplier.add()
 
 @app.route("/supplier/read/<id>/")
 def supplier_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('supplier/detail.html')
     # return Supplier.read(id)
 
 @app.route("/supplier/update/")
 def supplier_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Supplier.update()
 
 @app.route("/supplier/delete/")
 def supplier_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Supplier.delete()
 
 # End Supplier Route
@@ -383,23 +494,38 @@ def supplier_delete():
 
 @app.route("/customer/")
 def customer():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('customer/index.html')
 
 @app.route("/customer/add/")
 def customer_add():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Customer.add()
 
 @app.route("/customer/read/<id>/")
 def customer_read(id=None):
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return render_template('customer/detail.html')
     # return Customer.read(id)
 
 @app.route("/customer/update/")
 def customer_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Customer.update()
 
 @app.route("/customer/delete/")
 def customer_delete():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return Customer.delete()
 
 # End Customer Route
@@ -407,27 +533,42 @@ def customer_delete():
 
 @app.route("/user/")
 def user():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     return User.read()
 
 @app.route("/user/login/", methods=['POST'])
 def user_login():
     user = escape(request.form['username'])
     password = request.form['password']
+
     return User.login(user,password)
 
 @app.route("/user/add/", methods = ["POST"])
 def user_add():
     user = escape(request.form['username'])
     password = request.form['password']
+
+    if len(password) < 8:
+        flash('Please use longer password (Minimum 8 characters)')
+        return redirect(url_for('login'))
+
     return User.add(user,password)
 
 @app.route("/user/read/", methods = ["POST"])
 def user_read():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     id = escape(request.form['search_id']) if request.form['search_id'] else None
     return User.read(id)
 
 @app.route("/user/update/", methods = ["POST"])
 def user_update():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
     module = ['AP','AR','IV','PR','US']
     user_id = escape(request.form['US_ID'])
     username = escape(request.form['US_NAME'])
@@ -443,9 +584,32 @@ def user_update():
 
     return User.update(user_id, username, module_permission)
 
-@app.route("/user/delete/")
-def user_delete():
-    return User.delete()
+@app.route("/user/update/password/", methods = ["POST"])
+def user_update_password():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
+    id = request.form['US_ID']
+    password = request.form['US_PASSWORD']
+
+    return User.update(id,None,None,password)
+
+
+@app.route("/user/deactivate/", methods = ["POST"])
+def user_deactivate():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
+    user_id = escape(request.form['US_ID'])
+    return User.deactivate(user_id)
+
+@app.route("/user/activate/", methods = ["POST"])
+def user_activate():
+    if not User.is_logged_in():
+        return redirect(url_for('login'))
+
+    user_id = escape(request.form['US_ID'])
+    return User.activate(user_id)
 
 # End User Route
